include version of langchain and openai
